requirements.txt
Flask==3.0.3
groq==0.9.0
python-dotenv==1.0.1
gunicorn
flask-cors